Moved the variety query method to its own file
