+++ server/schemas/queries/index.js
@@ -1,0 +1,2 @@
+export { getAllUsers, getSpecificUser, getMe } from "./user.js";
+export { getVarieties } from "./variety.js";
